DynamicalSystems classes review : fix impact on OSI, fix tests, fix (some) examples- Try to merge with origin.

+++ examples/Electronics/Circuit-ThetaGammaSchemes/DiodeBridgeCapFilter-ThetaGammaSchemes.py
@@ -164,7 +164,7 @@ k = 0
 h = aTS.timeStep()
 print("Timestep : ", h)
 # Number of time steps
-N = (T - t0) / h
+N = int((T - t0) / h)
 print("Number of steps : ", N)
 
 # Get the values to be plotted
